+++ src/lib.rs
@@ -1,5 +1,6 @@
 pub mod cli;
 pub mod op;
+pub mod units;
 pub mod util;
 
 use std::error::Error;
